rename: CategoryPage용 폴더
