+++ src/components/CategoryPage/MonsterCard/MonsterCard.tsx
@@ -1,7 +1,7 @@
 import { MonsterName } from '@/model/monster';
 import Image from 'next/image';
 import styles from './css/MonsterCard.module.css';
-import BookMarkStarIcon from '../icons/BookMarkStarIcon';
+import BookMarkStarIcon from '../../icons/BookMarkStarIcon';
 
 type Props = {
   monster: MonsterName;
