+++ src/components/CategoryPage/RegisterByCategory/css/RegisterByCategory.module.css
