+++ src/components/CategoryPage/Registration/Label.tsx
