+++ src/components/CategoryPage/Registration/css/Descriptions.module.css
