+++ src/components/CategoryPage/Registration/css/Label.module.css
