+++ src/components/CategoryPage/Registration/css/Location.module.css
