+++ src/components/CategoryPage/Registration/css/Registration.module.css
