+++ src/components/CategoryPage/Registration/css/TimerInfo.module.css
